Refactor Game loop into helper methods
